Move Filter parse tests to own file
